matrix multiply
power function fixed by initialising with ones and using normal convention for counting index. Multiply matrix working but not many checks in place for bad inputs.
Discovered clone() so that we don't reference the same array. Forgot that the function gave an output in main. Still considering whether line 12 in main is an appropriate way to give an output into a new array to read out mathematically, but this works for now. Considered initialising the matrix on startup but i suppose this is unecessary. just an extra line that might be forgotten, its just annoying having the 'matrix not created' print for each function.