Triangle Interior Complete Untested
I did some reading on how the convex hull works, and decided to write my own that was optimised for the case where we are only dealing triangles. Normal algorithms such as gift wrapping considers the angle of each point and will survey every point, however, I can check only a few angles to get to the answer faster. Main feature of my algorithm is the rejection of non-intersecting query points before calculating the next test case, so that, assuming most rays don't intersect the triangle, no time is wasted. I start with a course check of a box, and as I go finer, the box gives me a nice boundary to work within, so only two angles need to be calculated.

I created the relevant methods in the matrix class that I would need to make it work. Everything is untested.

I have started using overloading. I may want to consider overloading the matrix initialisation in the future if I want other ways of generating them.

I realise functions like magnitude of a vector might be best pulled from the math library in java, but I am still learning what is available.