Create new basis complete
Test included.
Algorithm included in png. COB is P inv for vector v.

Matrix class now has overloading and all calls within matrix methods have been adjusted. specifying "empty" was pointless, and the IDE said that the vararg inputs for new inputs were confusing.

I needed to correct the following:
multiply by method returns a matrix as per the convention for all the others methods.
transpose works for rectangular matrices so that I flip the columns by the rows.
Counting in the index column and row methods was adjusted since the size of the output is different to the size of the input much of the time, so the index of the output was only adjusted after each output assignment.

Created indexRow method for boolean input array and single row extraction at index.

I made dot product accommodate row and column vectors, for any input.

Created an inverse method to work on a 3 by 3 matrix. I tested this for a random matrix A, multiplying A with inv(A) gave the identity.

Created determinant method for 2 by 2 matrix.

created a normalise columns method, but instead of returning the normals, which I don't need yet, or foreseeably need, it returns the normalised columns instead.