Trace Skeleton
Wrote out the rough skeleton for how the trace loop works, with some placeholder methods to be filled in shortly. For loop is safer than while loop because I can put in loop limit, then use the if statement like a while loop as long as the loop limit isn't too big, but it will just cycle through to the limit doing nothing each time after rays have all been blocked, which is checked by the raysareactive method.

trace step executes a single journey of each ray from one intersection to the next.

Created booleanArray so that I could give it a method to check row is false. to do it properly, I should have created an interface for the matrix, then have different types. Anyway this will suffice for now.

trace save in python is now update, which is a method belonging to the rays class now since there is nothing special about the triangulated shape that allows this to be written this way, also there are more references to rays than to the shape within the function.