distance line plane
Decided to process each ray-triangle pair within for loop that contains the index of each, and take each case as a single case by case analysis rather than bundling the results into vectors which was confusing.
There will be some functions I have created that won't be used because of this shift but are still there.

An interesting observation is that there are two different distance matrices, one for triangles and one for shapes. And they are treated differently in different contexts. Something to keep in mind.