trace distance complete. Update rays
I realised that the process for finding the shortest distance within a shape is the same between shapes so I re-arranged the larger of the two distance arrays so that rays span the rows, like the smaller. Then I could use the same process twice. Instead of grabbing the row I cared about, I still needed information about the triangle or the shape (depending on the context) so I used minRowIndex to give me information in both dimensions. Trace distance then outputs the normals of the shapes that I care about. Notice the 3 dimensional matrix is handled well as an array. Passing both the normals and the distances to the scene trace step allows me to update the rays ray by ray, simplifying the whole gig. I tried to minimise the number of inputs going into the update method. Didn't need to pass in shape, just the normal and the refractive index, in addition to the scene's refractive index.

minRow was important for maintaining information in 2 dimensions while keeping speed.

Rays combine sources became a constructor.

I now initialise indise and accumulated array lists.

Mirror is now refractive index of zero because this is an infinity light velocity therefore impossible.

I don't think I can fill a 2D array so in boolean array I fill each row instead.

Next step is reflect and refract. Then testing.